pyproject.toml
[tool.poetry]
name = "xes-conv-tcc"
version = "0.1.0"
description = ""
authors = ["Gustavo Albino <gustavo_albino@id.uff.br>"]
readme = "README.md"
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi-slim = "^0.111.0"
fastapi-injector = "^0.5.4"
gunicorn = "^22.0.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
aiohttp = "^3.9.5"
pydantic = "^2.7.1"
google-cloud-storage = "^2.16.0"
python-dateutil = "^2.9.0.post0"
types-google-cloud-ndb = "^2.3.0.20240311"
types-python-dateutil = "^2.9.0.20240316"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = 'py311'
exclude = [
    ".pytest_cache",
    ".mypy_cache",
    "__pypackages__",
    "__pycache__",
    ".git"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "PL",
    "UP",
    "SIM",
]
ignore = ["PLR2004", "PLR0913", "B008", "C408"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
